requirements.txt
requests
matplotlib
numpy
datetime
plotly
pandas
dash-bootstrap-components
emoji-country-flag
pycountry
ripe
ripe.atlas.cousteau
prettytable
node
tzdata
